feat(navigation): add menu API
