+++ tests/helpers/navMenuResponsive.test.js
@@ -1,14 +1,14 @@
 import { describe, it, expect, beforeEach } from "vitest";
-import { setupHamburgerMenu } from "../../src/helpers/navigation/navMenu.js";
+import { setupHamburger } from "../../src/helpers/api/navigation.js";
 
-describe("setupHamburgerMenu", () => {
+describe("setupHamburger", () => {
   beforeEach(() => {
     document.body.innerHTML = '<nav class="bottom-navbar"><ul><li></li></ul></nav>';
     window.innerWidth = 320;
   });
 
   it("creates a toggle button and toggles aria-expanded", () => {
-    setupHamburgerMenu();
+    setupHamburger();
     const button = document.querySelector(".nav-toggle");
     const list = document.querySelector(".bottom-navbar ul");
     expect(button).toBeTruthy();
@@ -19,7 +19,7 @@ describe("setupHamburgerMenu", () => {
   });
 
   it("removes the button when resized above breakpoint", () => {
-    setupHamburgerMenu();
+    setupHamburger();
     window.innerWidth = 1024;
     window.dispatchEvent(new Event("resize"));
     expect(document.querySelector(".nav-toggle")).toBeNull();
